move adapter methods GeminiAdaptersX
