+++ xchange-stream-gemini/src/main/java/info/bitrich/xchangestream/gemini/GeminiStreamingMarketDataService.java
@@ -98,7 +98,7 @@ public class GeminiStreamingMarketDataService implements StreamingMarketDataServ
                 });
 
     return subscribedOrderbookSnapshot.map(
-        geminiOrderbook -> geminiOrderbook.toOrderbook(maxDepth, new Date()));
+        geminiOrderbook -> GeminiAdaptersX.toOrderbook(geminiOrderbook, maxDepth, new Date()));
   }
 
   @Override
